third_party/WebKit/LayoutTests/http/tests/htmlimports/redirect-cross-origin-cross.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link rel="help" "https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html#imported-document" />

<link id="cors-xs" rel="import" href="http://localhost:8000/security/resources/cors-redir.php?url=http://localhost:8000/htmlimports/resources/hello.html">
</head>

<script>
function assert_loaded(id)
{
    var target = document.getElementById(id);
    assert_true(target.import instanceof Document)
    assert_true(0 <= target.import.querySelector("h1").innerHTML.indexOf("Hello"));
}

function assert_not_loaded(id)
{
    var target = document.getElementById(id);
    assert_equals(target.import, null);
}

test(function() { assert_not_loaded("cors-xs"); }, "Blocking access to a cross origin import including CORS domain redirect");
</script>
</body>
</html>
