third_party/WebKit/LayoutTests/http/tests/htmlimports/redirect-cross-origin-cross-expected.txt

CONSOLE ERROR: Imported resource from origin 'http://localhost:8000' has been blocked from loading by Cross-Origin Resource Sharing policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS Blocking access to a cross origin import including CORS domain redirect 
Harness: the test ran to completion.
